backend/requirements.txt
pandas==2.0.0
openpyxl==3.1.2
langchain==0.0.267
langchain-community==0.3.23
faiss-cpu==1.7.4
sentence-transformers==2.2.2
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==1.10.8
python-dotenv==1.0.0